Reference-count identifier strings (#1)
This way, we don't have to copy the entire text of the string when
returning the `IndexMap`. Since `Arc<str>` implements `Borrow<str>`,
the index map can still be accessed using `&str`s or `String`s as keys.

As a follow-up, now that `Element`s are cheap to clone, we could
consider changing `matches` to return an error that indicates the
expected element at the position that didn't match. But, this may not
be all that useful.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>