Add support for named suffix globs
Suffix globs cannot be captured to be included in the result map. This
change updates the pathscheme syntax to support named globs, in the form
`/src/*path`, to match suffix paths by-name. If the glob does not match
any path parts (for instance, when matching `/src`), then the match is
omitted from the result map.

This syntax is inspired by `httprouter`'s syntax.